The src attribute of html.
The <img> tag is used to embed an image in an HTML page. The src attribute specifies the path to the image to be displayed. There are two ways to specify the URL in the src attribute:
1. Absolute URL - Links to an external image that is hosted on another website. Example: src="https://www.w3schools.com/images/img_girl.jpg".
Notes: External images might be under copyright. If you do not get permission to use it, you may be in violation of copyright laws. In addition, you cannot control external images; it can suddenly be removed or changed.
2. Relative URL - Links to an image that is hosted within the website. Here, the URL does not include the domain name. If the URL begins without a slash, it will be relative to the current page. Example: src="img_girl.jpg". If the URL begins with a slash, it will be relative to the domain. Example: src="/images/img_girl.jpg".

Note: It is almost always best to use relative URLs. They will not break if you change domain.

The <img> tag should also contain the width and height attributes, which specifies the width and height of the image (in pixels).