Use of quotations in html.
The HTML <blockquote> element defines a section that is quoted from another source. Browsers usually indent <blockquote> elements.
The HTML <q> tag defines a short quotation. Browsers normally insert quotation marks around the quotation.
The HTML <abbr> tag defines an abbreviation or an acronym, like "HTML", "CSS", "Mr.", "Dr.", "ASAP", "ATM". Marking abbreviations can give useful information to browsers, translation systems and search-engines.
Tip: Use the global title attribute to show the description for the abbreviation/acronym when you mouse over the element.
The HTML <address> tag defines the contact information for the author/owner of a document or an article. The contact information can be an email address, URL, physical address, phone number, social media handle, etc. The text in the <address> element usually renders in italic, and browsers will always add a line break before and after the <address> element.
The HTML <cite> tag defines the title of creative work (e.g. a book, a poem, a song, a movie, a painting, a sculpture, etc.).
Note: A person's name is not the title of a work.
The text in the <cite> element usually renders in italic.
BDO stands for Bi-Directional Override. The HTML <bdo> tag is used to override the current text direction.