Comments and colors in html.
HTML comments are not displayed in the browser, but they can help document your HTML source code.You can add comments to your HTML source by using the following syntax. Notice that there is an exclamation point (!) in the start tag, but not in the end tag.
Note: Comments are not displayed by the browser, but they can help document your HTML source code. With comments, you can place notifications and reminders in your HTML code.Comments are also great for debugging HTML, because you can comment out HTML lines of code, one at a time, to search for errors.