Internal & External CSS in html.
An internal CSS is used to define a style for a single HTML page.
An internal CSS is defined in the <head> section of an HTML page, within a <style> element.
The following example sets the text color of ALL the <h1> elements (on that page) to blue and the text color of ALL the <p> elements to red. In addition, the page will be displayed with a "powderblue" background color.
An external style sheet is used to define the style for many HTML pages. To use an external style sheet, add a link to it in the <head> section of each HTML page. The external style sheet can be written in any text editor. The file must not contain any HTML code and must be saved with a .css extension.
Tip: With an external style sheet, you can change the look of an entire website, by changing one file!