Tables in html.
A table in HTML consists of table cells inside rows and columns.
Table Cells
Each table cell is defined by a <td> and a </td> tag.
td stands for table data. Everything between <td> and </td> are the content of the table cell.
Table Rows
Each table row starts with a <tr> and end with a </tr> tag. tr stands for table row. You can have as many rows as you like in a table, just make sure that the number of cells are the same in each row.
Table Headers
Sometimes you want your cells to be headers, in those cases use the <th> tag instead of the <td> tag.